Remove the `model` suffix on the `support.workers` package name
