+++ src/main/scala/com/gu/support/workers/AcquisitionData.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model
+package com.gu.support.workers
 
 import com.gu.acquisition.model.{OphanIds, ReferrerAcquisitionData}
 import ophan.thrift.event.AbTest
