+++ src/main/scala/com/gu/support/workers/BillingPeriod.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model
+package com.gu.support.workers
 
 import io.circe.{Decoder, Encoder}
