+++ src/main/scala/com/gu/support/workers/ExecutionError.scala
@@ -1,3 +1,3 @@
-package com.gu.support.workers.model
+package com.gu.support.workers
 
 case class ExecutionError(Error: String, Cause: String)
