+++ src/main/scala/com/gu/support/workers/Products.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model
+package com.gu.support.workers
 
 import com.gu.i18n.Currency
 import com.gu.support.encoding.Codec
