+++ src/main/scala/com/gu/support/workers/RequestInfo.scala
@@ -1,4 +1,4 @@
-package com.gu.support.workers.model
+package com.gu.support.workers
 
 case class RequestInfo(encrypted: Boolean, testUser: Boolean, failed: Boolean, messages: List[String]){
   def appendMessage(message: String) = copy(messages = messages :+ message)
