+++ src/main/scala/com/gu/support/workers/SalesforceContactRecord.scala
@@ -1,0 +1,3 @@
+package com.gu.support.workers
+
+case class SalesforceContactRecord(Id: String, AccountId: String)
