--- src/main/scala/com/gu/support/workers/model/SalesforceContactRecord.scala
@@ -1,3 +1,0 @@
-package com.gu.support.workers.model
-
-case class SalesforceContactRecord(Id: String, AccountId: String)
