--- src/main/scala/com/gu/support/workers/model/states/CheckoutFailureState.scala
@@ -1,8 +1,0 @@
-package com.gu.support.workers.model.states
-
-import com.gu.support.workers.model.CheckoutFailureReasons.CheckoutFailureReason
-import com.gu.support.workers.model.User
-
-case class CheckoutFailureState(user: User, checkoutFailureReason: CheckoutFailureReason) extends StepFunctionUserState
-
-
