--- src/main/scala/com/gu/support/workers/model/states/SendThankYouEmailState.scala
@@ -1,16 +1,0 @@
-package com.gu.support.workers.model.states
-
-import java.util.UUID
-
-import com.gu.support.workers.model._
-
-case class SendThankYouEmailState(
-  requestId: UUID,
-  user: User,
-  product: ProductType,
-  paymentMethod: PaymentMethod,
-  salesForceContact: SalesforceContactRecord,
-  accountNumber: String,
-  acquisitionData: Option[AcquisitionData]
-) extends StepFunctionUserState
-
