--- src/main/scala/com/gu/support/workers/model/states/StepFunctionUserState.scala
@@ -1,7 +1,0 @@
-package com.gu.support.workers.model.states
-
-import com.gu.support.workers.model.User
-
-trait StepFunctionUserState {
-  def user: User
-}
