+++ src/main/scala/com/gu/support/workers/states/CheckoutFailureState.scala
@@ -1,0 +1,8 @@
+package com.gu.support.workers.states
+
+import com.gu.support.workers.User
+import com.gu.support.workers.CheckoutFailureReasons.CheckoutFailureReason
+
+case class CheckoutFailureState(user: User, checkoutFailureReason: CheckoutFailureReason) extends StepFunctionUserState
+
+
