+++ src/main/scala/com/gu/support/workers/states/CreateSalesforceContactState.scala
@@ -1,16 +1,16 @@
-package com.gu.support.workers.model.states
+package com.gu.support.workers.states
 
 import java.util.UUID
 
 import com.gu.support.promotions.PromoCode
-import com.gu.support.workers.model._
+import com.gu.support.workers.{PaymentMethod, User}
+import com.gu.support.workers._
 
-case class CreatePaymentMethodState(
+case class CreateSalesforceContactState(
   requestId: UUID,
   user: User,
   product: ProductType,
-  paymentFields: PaymentFields,
+  paymentMethod: PaymentMethod,
   promoCode: Option[PromoCode],
   acquisitionData: Option[AcquisitionData]
 ) extends StepFunctionUserState
-
