+++ src/main/scala/com/gu/support/workers/states/FailureHandlerState.scala
@@ -1,8 +1,9 @@
-package com.gu.support.workers.model.states
+package com.gu.support.workers.states
 
 import java.util.UUID
 
-import com.gu.support.workers.model._
+import com.gu.support.workers.User
+import com.gu.support.workers._
 
 case class FailureHandlerState(
   requestId: UUID,
