+++ src/main/scala/com/gu/support/workers/states/StepFunctionUserState.scala
@@ -1,0 +1,7 @@
+package com.gu.support.workers.states
+
+import com.gu.support.workers.User
+
+trait StepFunctionUserState {
+  def user: User
+}
