+++ src/main/scala/com/gu/support/zuora/api/PaymentGateway.scala
@@ -2,7 +2,8 @@ package com.gu.support.zuora.api
 
 import com.gu.i18n.Currency
 import com.gu.i18n.Currency.AUD
-import com.gu.support.workers.model._
+import com.gu.support.workers.{CreditCardReferenceTransaction, DirectDebitPaymentMethod, PayPalReferenceTransaction, PaymentMethod}
+import com.gu.support.workers._
 import io.circe.{Decoder, Encoder}
 
 import scala.PartialFunction.condOpt
