+++ src/test/scala/com/gu/support/zuora/api/Fixtures.scala
@@ -2,7 +2,8 @@ package com.gu.support.zuora.api
 
 import com.gu.i18n.Currency.GBP
 import com.gu.i18n.{Country, Currency}
-import com.gu.support.workers.model.{CreditCardReferenceTransaction, DirectDebitPaymentMethod, PayPalReferenceTransaction}
+import com.gu.support.workers.{CreditCardReferenceTransaction, DirectDebitPaymentMethod, PayPalReferenceTransaction}
+import com.gu.support.workers.{DirectDebitPaymentMethod, PayPalReferenceTransaction}
 import org.joda.time.LocalDate
 
 //noinspection TypeAnnotation
