+++ src/test/scala/com/gu/support/zuora/api/SerialisationSpec.scala
@@ -4,7 +4,7 @@ import com.gu.i18n.Currency.GBP
 import com.gu.support.SerialisationTestHelpers
 import com.gu.support.SerialisationTestHelpers._
 import com.gu.support.encoding.CustomCodecs._
-import com.gu.support.workers.model.PaymentFields
+import com.gu.support.workers.PaymentFields
 import com.gu.support.zuora.api.Fixtures._
 import com.gu.support.zuora.api.response._
 import com.typesafe.scalalogging.LazyLogging
